# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GovCore Platform API — Identity Engine
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents
      parameters:
        - name: tenantId
          in: query
          schema:
            type: integer
        - name: module
          in: query
          schema:
            type: string
        - name: documentType
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, generated, reviewed, approved, signed, archived, retained, disposed]
        - name: resourceType
          in: query
          schema:
            type: string
        - name: resourceId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Create a document, optionally generating its first version immediately
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Document created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentDetail"
        "400":
          description: Template variables missing
        "404":
          description: Template not found
  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get a document with its versions, signatures, and attachments
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: The document and its detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentDetail"
        "404":
          description: Document not found
  /documents/{id}/generate:
    post:
      operationId: generateDocument
      tags: [documents]
      summary: Render the document's template into a new immutable version
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentGenerateInput"
      responses:
        "200":
          description: New version generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentDetail"
        "400":
          description: Template variables missing
        "404":
          description: Document or template not found
  /documents/{id}/status:
    post:
      operationId: updateDocumentStatus
      tags: [documents]
      summary: Advance a document through its lifecycle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentStatusInput"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Document not found
        "409":
          description: Lifecycle stages only move forward
  /documents/{id}/content:
    get:
      operationId: getDocumentContent
      tags: [documents]
      summary: Download the rendered content of a document version, recording the access
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rendered content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentContent"
        "404":
          description: Document or version not found
  /documents/{id}/sign:
    post:
      operationId: signDocument
      tags: [documents]
      summary: Apply an electronic signature to the document's current version
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentSignatureInput"
      responses:
        "201":
          description: Signature recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentSignature"
        "404":
          description: Document not found
        "409":
          description: Document has no generated version to sign
  /documents/{id}/access-logs:
    get:
      operationId: listDocumentAccessLogs
      tags: [documents]
      summary: Access and lifecycle trail for a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Access events, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DocumentAccessLog"
  /documents/verify/{uuid}:
    get:
      operationId: verifyDocument
      tags: [documents]
      summary: Public QR verification — confirms a document was issued without revealing its contents
      security: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentVerification"
...
components:
  schemas:
...
    Document:
      type: object
      required: [id, uuid, tenantId, title, documentType, module, status, currentVersion, createdAt, updatedAt]
      properties:
        id:
          type: integer
        uuid:
          type: string
          description: Public identifier used by storage keys and QR verification
        tenantId:
          type: integer
        templateId:
          type: integer
          nullable: true
        title:
          type: string
        documentType:
          type: string
        module:
          type: string
        resourceType:
          type: string
          nullable: true
        resourceId:
          type: string
          nullable: true
        referenceNumber:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, generated, reviewed, approved, signed, archived, retained, disposed]
        currentVersion:
          type: integer
          description: 0 until the document has been generated
        contentHash:
          type: string
          nullable: true
        retainUntil:
          type: string
          format: date-time
          nullable: true
        createdByUserId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DocumentInput:
      type: object
      required: [tenantId, title, module]
      properties:
        tenantId:
          type: integer
        title:
          type: string
        module:
          type: string
        documentType:
          type: string
        templateCode:
          type: string
          description: Resolves the active template for this code/locale
        locale:
          type: string
        resourceType:
          type: string
        resourceId:
          type: string
        payload:
          type: object
          description: Variable values; supplying this generates version 1 immediately
          additionalProperties: true
        generate:
          type: boolean
          description: Generate the first version on creation (default true when templateCode and payload are supplied)
    DocumentGenerateInput:
      type: object
      required: [payload]
      properties:
        templateCode:
          type: string
          description: Overrides the document's existing template
        locale:
          type: string
        payload:
          type: object
          additionalProperties: true
    DocumentStatusInput:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [draft, generated, reviewed, approved, signed, archived, retained, disposed]
        detail:
          type: string
        retainUntil:
          type: string
          format: date-time
    DocumentVersion:
      type: object
      required: [id, documentId, version, mimeType, storageProvider, contentHash, createdAt]
      properties:
        id:
          type: integer
        documentId:
          type: integer
        version:
          type: integer
        mimeType:
          type: string
        content:
          type: string
          nullable: true
        storageProvider:
          type: string
          enum: [inline, local, nas, s3]
        storageKey:
          type: string
          nullable: true
        sizeBytes:
          type: integer
          nullable: true
        contentHash:
          type: string
        payload:
          type: string
          nullable: true
        generatedByUserId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
    DocumentContent:
      type: object
      required: [documentId, version, mimeType, content, contentHash]
      properties:
        documentId:
          type: integer
        version:
          type: integer
        mimeType:
          type: string
        content:
          type: string
        contentHash:
          type: string
    DocumentSignature:
      type: object
      required: [id, documentId, documentVersionId, tenantId, signerName, signatureType, signedHash, signedAt]
      properties:
        id:
          type: integer
        documentId:
          type: integer
        documentVersionId:
          type: integer
        tenantId:
          type: integer
        signerUserId:
          type: integer
          nullable: true
        signerName:
          type: string
        signerRole:
          type: string
          nullable: true
        signatureType:
          type: string
          enum: [electronic, digital]
        signatureData:
          type: string
          nullable: true
        signedHash:
          type: string
        certificateSubject:
          type: string
          nullable: true
        signedAt:
          type: string
          format: date-time
    DocumentSignatureInput:
      type: object
      required: [signerName]
      properties:
        signerName:
          type: string
        signerRole:
          type: string
        signatureData:
          type: string
          description: Data URI of a drawn signature, when captured
    DocumentAccessLog:
      type: object
      required: [id, documentId, tenantId, action, createdAt]
      properties:
        id:
          type: integer
        documentId:
          type: integer
        documentVersionId:
          type: integer
          nullable: true
        tenantId:
          type: integer
        action:
          type: string
        detail:
          type: string
          nullable: true
        actorUserId:
          type: integer
          nullable: true
        ipAddress:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    DocumentDetail:
      allOf:
        - $ref: "#/components/schemas/Document"
        - type: object
          required: [versions, signatures, attachments]
          properties:
            versions:
              type: array
              items:
                $ref: "#/components/schemas/DocumentVersion"
            signatures:
              type: array
              items:
                $ref: "#/components/schemas/DocumentSignature"
            attachments:
              type: array
              items:
                $ref: "#/components/schemas/Attachment"
            missingVariables:
              type: array
              description: Present when a generation left placeholders unresolved
              items:
                type: string
    DocumentVerification:
      type: object
      required: [valid]
      properties:
        valid:
          type: boolean
        reason:
          type: string
        document:
          type: object
          properties:
            uuid:
              type: string
            title:
              type: string
            documentType:
              type: string
            referenceNumber:
              type: string
              nullable: true
            status:
              type: string
            version:
              type: integer
            issuedAt:
              type: string
              format: date-time
        signatures:
          type: array
          items:
            type: object
            properties:
              signerName:
                type: string
              signerRole:
                type: string
                nullable: true
              signedAt:
                type: string
                format: date-time
    Attachment:
      type: object
      required: [id, uuid, tenantId, fileName, mimeType, sizeBytes, storageProvider, createdAt]
      properties:
        id:
          type: integer
        uuid:
          type: string
          description: The value a form file field stores in its submission
        tenantId:
          type: integer
        documentId:
          type: integer
          nullable: true
        attachedToType:
          type: string
          nullable: true
        attachedToId:
          type: string
          nullable: true
        fieldKey:
          type: string
          nullable: true
        fileName:
          type: string
        mimeType:
          type: string
        sizeBytes:
          type: integer
        storageProvider:
          type: string
        storageKey:
          type: string
          nullable: true
        contentHash:
          type: string
          nullable: true
        uploadedByUserId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
...